backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "psutil>=7.2.1",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-crontab>=3.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "redis>=7.1.0",
    "ruff>=0.14.14",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B", "A", "C4", "PT", "RUF"]
ignore = []

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
